feat(discord-bot): make the Worker bundleable — split the discord barrel
The Worker now builds and boots on workerd. 678 KiB raw, 115 KiB gzipped —
comfortably inside the 3 MB free-tier limit — with no gateway client in the
bundle. Verified with `wrangler deploy --dry-run` and `wrangler dev`.

`utils/discord.ts` was doing two unrelated jobs: handing out builders and enums,
which are pure data construction, and handing out `Client`/`Events`/
`GatewayIntentBits`, which exist only to hold a WebSocket. Command files needed
the first and got the second, so every command transitively imported discord.js
and no Worker build could ever succeed.

`utils/builders.ts` is now the portable half, and the rule is: command files
import from there, never from `discord.ts`. Anything reaching for the gateway
barrel is transport and belongs in the gateway entry point.

Three things this turned up that the plan did not anticipate:

**discord.js does not re-export @discordjs/builders — it SUBCLASSES it.** Its
`EmbedBuilder` adds `.length` and accepts hex-string colours, so the two are
genuinely distinct types and a command returning one cannot satisfy a signature
demanding the other. "Just import from the underlying package" is not the
drop-in it sounds like. The portable builder wins here because command renderers
have to run on workerd, and the gateway path accepts it happily since it is a
JSONEncodable<APIEmbed>. Five hex-string `setColor` calls became numeric
literals, which both builders accept.

**`errorTracker` imported `randomUUID` from `node:crypto`**, which made the
module unloadable in a Worker outright. Switched to the WebCrypto global:
identical output, exists on both runtimes, and avoids pulling in nodejs_compat
and a pile of Node polyfills for the sake of one UUID.

**Type-only imports of gateway types are fine and stay.** Three files still
`import type` from `discord.ts`; those are erased at build time and never reach
the bundle. Confirmed by inspecting the output — no `WebSocketManager`, no
discord.js client. The only gateway-shaped string left is `GatewayIntentBits`,
which is a plain enum from discord-api-types.

Also adds `apps/discord-bot/wrangler.jsonc`, deliberately NOT wired to any
deploy. Deploying it changes nothing on its own: Discord only routes here once
an Interactions Endpoint URL is set, and setting that stops gateway delivery
entirely. `DISCORD_PUBLIC_KEY` is left unset, so the Worker currently rejects
every request — the correct direction to fail in.

Local verification: unsigned POST 401, bad-signature POST 401, GET 405.

Still not at parity — /help, /notation and /salvageunion have no Worker renderer
yet, and /notation's collector-based pagination still needs redesigning around
custom_id. Do not point Discord at this.
