+++ apps/discord-bot/package.json
@@ -16,8 +16,10 @@
   },
   "type": "module",
   "dependencies": {
+    "@discordjs/builders": "1.14.1",
     "@randsum/games": "workspace:~",
     "@randsum/roller": "workspace:~",
+    "discord-api-types": "0.38.49",
     "discord.js": "^14.26.4"
   },
   "keywords": [
