+++ apps/discord-bot/src/utils/builders.ts
@@ -1,0 +1,33 @@
+/**
+ * Portable Discord primitives — safe on Node and on workerd.
+ *
+ * The counterpart to `discord.ts`, which is the *gateway* barrel: `Client`,
+ * `Collection`, `Events`, `GatewayIntentBits`. Those exist only to hold and
+ * service a WebSocket, and none of them can run on Workers.
+ *
+ * Everything here is different in kind. Builders and enums are pure data
+ * construction with no runtime dependency on a connection — and crucially they
+ * do not actually live in discord.js at all. discord.js re-exports them from
+ * `@discordjs/builders` and `discord-api-types`, which are ordinary portable
+ * packages. Importing from the source rather than through discord.js is what
+ * lets a command file compile for both targets with no aliasing, no bundler
+ * tricks, and no second implementation.
+ *
+ * The version is pinned to the one discord.js itself depends on, so the classes
+ * here are the same classes it would hand back. A drift there would produce
+ * two `EmbedBuilder` identities in one process, which fails in a genuinely
+ * confusing way.
+ *
+ * The rule this encodes: **command files import from here, never from
+ * `discord.ts`.** Anything that needs the gateway barrel is transport, and
+ * belongs in the gateway entry point.
+ */
+export {
+  ActionRowBuilder,
+  ButtonBuilder,
+  EmbedBuilder,
+  SlashCommandBuilder,
+  StringSelectMenuBuilder
+} from '@discordjs/builders'
+
+export { ComponentType, MessageFlags } from 'discord-api-types/v10'
