+++ apps/discord-bot/wrangler.jsonc
@@ -1,0 +1,26 @@
+{
+  // Cloudflare Worker for the Discord bot — HTTP interactions.
+  //
+  // NOT WIRED UP. Deploying this changes nothing on its own: Discord only sends
+  // interactions here once an Interactions Endpoint URL is set in the
+  // application settings, and doing that STOPS gateway delivery entirely. The
+  // two transports are mutually exclusive, so the switch is a deliberate,
+  // one-way-feeling cutover — see apps/discord-bot/CLAUDE.md before flipping it.
+  //
+  // Wrangler bundles the TypeScript entry directly; there is no separate build
+  // step, unlike the gateway bot which bunup-bundles for Node.
+  "$schema": "node_modules/wrangler/config-schema.json",
+  "name": "randsum-discord-bot",
+  "main": "./src/worker/index.ts",
+  // Matches the sites: newer than the locally installed wrangler's workerd
+  // supports would make `wrangler dev` refuse to start, and local testing is
+  // the only way to exercise this before Discord is pointed at it.
+  "compatibility_date": "2026-07-13",
+  "observability": {
+    "enabled": true
+  }
+  // DISCORD_PUBLIC_KEY is intentionally absent. It is not a secret — it is the
+  // application's public key — but it is environment-specific, so it is set via
+  // `wrangler secret put` / the dashboard rather than committed. With it unset
+  // the Worker rejects every request, which is the correct failure direction.
+}
